+++ src/main/java/bowling/domain/bowl/formatter/GutterNormalBowlFormatter.java
@@ -1,4 +1,6 @@
-package bowling.domain;
+package bowling.domain.bowl.formatter;
+
+import bowling.domain.bowl.NormalBowl;
 
 import static bowling.domain.NumberOfPins.MIN_NUMBER_OF_PINS;
