+++ src/main/java/bowling/domain/bowl/formatter/NoneNormalBowlFormatter.java
@@ -1,4 +1,6 @@
-package bowling.domain;
+package bowling.domain.bowl.formatter;
+
+import bowling.domain.bowl.NormalBowl;
 
 public class NoneNormalBowlFormatter implements NormalBowlFormatter {
